# api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/v1/annonces/{id}": {
      "get": {
        "tags": [
          "annonce-controller"
        ],
        "operationId": "getAnnonceById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Annonce"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "annonce-controller"
        ],
        "operationId": "updateAnnonce",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Annonce"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Annonce"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "annonce-controller"
        ],
        "operationId": "deleteEmployee",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/register": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "operationId": "registerUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "SUCCESS",
                    "USER_ALREADY_EXISTS",
                    "FAILURE"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/logout": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "operationId": "logUserOut",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "SUCCESS",
                    "USER_ALREADY_EXISTS",
                    "FAILURE"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/login": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "operationId": "loginUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "SUCCESS",
                    "USER_ALREADY_EXISTS",
                    "FAILURE"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/annonces": {
      "get": {
        "tags": [
          "annonce-controller"
        ],
        "operationId": "getAllAnnonces",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Annonce"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "annonce-controller"
        ],
        "operationId": "createAnnonce",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Annonce"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Annonce"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/all": {
      "delete": {
        "tags": [
          "user-controller"
        ],
        "operationId": "deleteUsers",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "SUCCESS",
                    "USER_ALREADY_EXISTS",
                    "FAILURE"
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Annonce": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "titre": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "image_url": {
            "type": "string"
          },
          "date_debut": {
            "type": "string",
            "format": "date-time"
          },
          "date_fin": {
            "type": "string",
            "format": "date-time"
          },
          "dateCreation": {
            "type": "string",
            "format": "date"
          },
          "user": {
            "$ref": "#/components/schemas/UserEntity"
          }
        }
      },
      "AnnonceEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "titre": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "image_url": {
            "type": "string"
          },
          "date_debut": {
            "type": "string",
            "format": "date-time"
          },
          "date_fin": {
            "type": "string",
            "format": "date-time"
          },
          "dateCreation": {
            "type": "string",
            "format": "date"
          },
          "user": {
            "$ref": "#/components/schemas/UserEntity"
          }
        }
      },
      "UserEntity": {
        "required": [
          "loggedIn",
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "loggedIn": {
            "type": "boolean"
          },
          "annonces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnnonceEntity"
            }
          }
        }
      }
    }
  }
}